// apps/ripscloud-web/src/api/api.ts
export interface RipsAdminApplicationDTOsCreateInvoiceResolutionRequestDto {
    'resolutionNumber'?: string;
    'prefix'?: string;
    'fromNumber'?: number;
    'toNumber'?: number;
    'validFrom'?: string;
    'validTo'?: string;
    'isActive'?: boolean;
    'environment'?: number;
    'technicalKey'?: string;
}
// ...
export interface RipsAdminApplicationDTOsInvoiceResolutionDto {
    'id'?: string;
    'resolutionNumber'?: string;
    'prefix'?: string;
    'nextNumber'?: number;
    'fromNumber'?: number;
    'toNumber'?: number;
    'validFrom'?: string;
    'validTo'?: string;
    'isActive'?: boolean;
    'environment'?: number;
    'technicalKey'?: string;
    'createdAt'?: string;
    'updatedAt'?: string;
}
// ...
export interface RipsAdminApplicationDTOsUpdateInvoiceResolutionRequestDto {
    'resolutionNumber'?: string;
    'prefix'?: string;
    'fromNumber'?: number;
    'toNumber'?: number;
    'validFrom'?: string;
    'validTo'?: string;
    'isActive'?: boolean;
    'environment'?: number;
    'technicalKey'?: string;
}